+++ gatherer/src/pandas/gatherer/core/GatherException.java
@@ -1,12 +1,18 @@
 package pandas.gatherer.core;
 
 public class GatherException extends Exception {
-    public GatherException(String message) {
+    private Integer exitStatus;
+
+    public GatherException(String message, int exitStatus) {
         super(message);
+        this.exitStatus = exitStatus;
     }
 
     public GatherException(String message, Throwable cause) {
         super(message, cause);
     }
 
+    public Integer getExitStatus() {
+        return exitStatus;
+    }
 }